move validation to unit test
